css-text/line-breaking/css-text-line-break-ba-other.html

<!DOCTYPE html>
<html  lang="en" >
<head>
<meta charset="utf-8"/>
<title>Line-breaking: Other</title>
<link rel="author" title="Richard Ishida" href="mailto:ishida@w3.org">
<link rel="help" href="https://drafts.csswg.org/css-text-3/#line-breaking">
<script src="../../resources/testharness.js"></script>
<script src="../../resources/testharnessreport.js"></script>
<meta name="assert" content="[Exploratory] The browser will break a line of text after each of the BA characters listed.">
<style type="text/css">
.breakable { font-size: 25px; width: 800px; line-height: 30px; }
</style>
<script>
var charlist = `1DA87 SIGNWRITING COMMA
1DA88 SIGNWRITING FULL STOP
1DA89 SIGNWRITING SEMICOLON
1DA8A SIGNWRITING COLON
2056 THREE DOT PUNCTUATION
2058 FOUR DOT PUNCTUATION
2059 FIVE DOT PUNCTUATION
205A TWO DOT PUNCTUATION
205B FOUR DOT MARK
205D TRICOLON
205E VERTICAL FOUR DOTS
2E0E EDITORIAL CORONIS
2E0F PARAGRAPHOS
2E10 FORKED PARAGRAPHOS
2E11 REVERSED FORKED PARAGRAPHOS
2E12 HYPODIASTOLE
2E13 DOTTED OBELOS
2E14 DOWNWARDS ANCORA
2E15 UPWARDS ANCORA
2E17 DOUBLE OBLIQUE HYPHEN
2E19 PALM BRANCH
2E2A TWO DOTS OVER ONE DOT PUNCTUATION
2E2B ONE DOT OVER TWO DOTS PUNCTUATION
2E2C SQUARED FOUR DOT PUNCTUATION
2E2D FIVE DOT MARK
2E30 RING POINT
2E31 WORD SEPARATOR MIDDLE DOT
2E33 RAISED DOT
2E34 RAISED COMMA
2E3C STENOGRAPHIC FULL STOP
2E3D VERTICAL SIX DOTS
2E3E WIGGLY VERTICAL LINE
2E40 DOUBLE HYPHEN
2E41 REVERSED COMMA
2E43 DASH WITH LEFT UPTURN
2E44 DOUBLE SUSPENSION MARK
2E45 INVERTED LOW KAVYKA
2E46 INVERTED LOW KAVYKA WITH KAVYKA ABOVE
2E47 LOW KAVYKA
2E48 LOW KAVYKA WITH DOT
2E49 DOUBLE STACKED COMMA
10B3A TINY TWO DOTS OVER ONE DOT PUNCTUATION
10B3B SMALL TWO DOTS OVER ONE DOT PUNCTUATION
10B3C LARGE TWO DOTS OVER ONE DOT PUNCTUATION
10B3D LARGE ONE DOT OVER TWO DOTS PUNCTUATION
10B3E LARGE TWO RINGS OVER ONE RING PUNCTUATION
10B3F LARGE ONE RING OVER TWO RINGS PUNCTUATION`
</script>
</head>
<body>
<script>
var lines = charlist.split('\n')
var out = '<div id="log"></div>\n'
for (var i=0;i<lines.length;i++) {
    // get the data
    var firstSpace = lines[i].indexOf(' ')
    var hex = lines[i].substr(0,firstSpace)
    var name = lines[i].substr(firstSpace)
    // make a test
    out +=  '<div><div>'+hex+'</div><div id="test'+i+'" class="breakable">aaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaaa&#x'+hex+';bbbbbbbbbbbbbbbbbbbbbbbbbbbbbb</div></div>'
    }
document.querySelector('body').innerHTML = out
// run the test framework
 for (i=0;i<lines.length;i++) {
   test(function() {
        assert_true(document.getElementById('test'+i).offsetHeight > 35);
        }, lines[i]+' should allow a line break after');
   }
// hide successful tests
for (i=0;i<lines.length;i++) {
    if (document.getElementById('test'+i).offsetHeight > 35) document.getElementById('test'+i).parentNode.style.display = 'none'
    }
</script>
<!--Notes:
This list is based on property assignments for Unicode version 10.
Each test creates an unbroken text item from two very long words connected by the character being tested. The whole text item is displayed in a box that is narrower than the width of the text item.  If the browser allows a line to wrap after the character being tested, the text item will break.  The test measures the height of the text box to determine whether the text wrapped.
-->
</body>
</html>
